Keep track of open terminals
